+++ CHANGES.md
@@ -37,7 +37,7 @@
 - Added new relation type: RL_SND_MSG_Q and RL_RCV_MSG_Q.
 - Added new relation type: RL_SND_MSG and RL_RCV_MSG.
 - Added new relation type: RL_CLONE_MEM.
-- Added new relation type: RL_SOCKET_CREATE.
+- Split RL_CREATE into RL_SOCKET_CREATE and RL_MSG_CREATE.
 - Re-introduce RL_SND_UNIX and RL_RCV_UNIX now that we have space.
 - Added new relation type: RL_PROC_WRITE and RL_PROC_READ.
 - Increased the maximum number of relation type supported.